chore: split test into `suite` builders
