+++ test/suites/index.js
@@ -1,0 +1,21 @@
+export { default as API } from './api';
+
+export { default as Numbers } from './number';
+export { default as Booleans } from './boolean';
+export { default as Strings } from './string';
+export { default as Nully } from './nullish';
+
+export { default as Maps } from './map';
+export { default as Sets } from './set';
+
+export { default as TypedArrays } from './typedarray';
+
+export { default as Dates } from './date';
+export { default as RegExps } from './regexp';
+
+export { default as Objects } from './object';
+export { default as Arrays } from './array';
+
+export { default as Functions } from './function';
+export { default as Pollutions } from './pollution';
+export { default as Classes } from './class';
